pyproject.toml
[tool.poetry]
name = "no-more-imagenet"
version = "0.1.0"
description = ""
authors = ["Harry Hsu <stmharry@hashgreen.net>"]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^1.13.1"
torchmetrics = "^0.11.0"
torchvision = "^0.14.1"
torchtyping = "^0.1.4"
absl-py = "^1.3.0"
pandas = "^1.5.2"
PyYAML = "^6.0"
pydantic = "^1.10.4"
Pillow = "^9.4.0"
rich = "^13.0.1"

[tool.poetry.dev-dependencies]
devtools = {extras = ["pygments"], version = "^0.10.0"}
mypy = "^0.991"
types-PyYAML = "^6.0.12"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
